requirements.txt
pillow
django-crispy-forms
crispy_bootstrap4
bootstrap4
djangorestframework